requirements.txt
langchain
langchain-core
langchain-google-genai>=2.1.9
langgraph>=0.6.3
pypdf2>=3.0.1
python-dotenv>=1.1.1
requests>=2.32.4
streamlit>=1.48.0
